# [3] graph/db: batch-fetch channel & policy data