--- a/sqldb/sqlc/graph.sql.go
+++ b/sqldb/sqlc/graph.sql.go
@@ -701,6 +701,49 @@ func (q *Queries) GetChannelBySCIDWithPolicies(ctx context.Context, arg GetChann
 	return i, err
 }
 
+const getChannelExtrasBatch = `-- name: GetChannelExtrasBatch :many
+SELECT
+    channel_id,
+    type,
+    value
+FROM graph_channel_extra_types
+WHERE channel_id IN (/*SLICE:chan_ids*/?)
+ORDER BY channel_id, type
+`
+
+func (q *Queries) GetChannelExtrasBatch(ctx context.Context, chanIds []int64) ([]GraphChannelExtraType, error) {
+	query := getChannelExtrasBatch
+	var queryParams []interface{}
+	if len(chanIds) > 0 {
+		for _, v := range chanIds {
+			queryParams = append(queryParams, v)
+		}
+		query = strings.Replace(query, "/*SLICE:chan_ids*/?", makeQueryParams(len(queryParams), len(chanIds)), 1)
+	} else {
+		query = strings.Replace(query, "/*SLICE:chan_ids*/?", "NULL", 1)
+	}
+	rows, err := q.db.QueryContext(ctx, query, queryParams...)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var items []GraphChannelExtraType
+	for rows.Next() {
+		var i GraphChannelExtraType
+		if err := rows.Scan(&i.ChannelID, &i.Type, &i.Value); err != nil {
+			return nil, err
+		}
+		items = append(items, i)
+	}
+	if err := rows.Close(); err != nil {
+		return nil, err
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return items, nil
+}
+
 const getChannelFeaturesAndExtras = `-- name: GetChannelFeaturesAndExtras :many
 SELECT
     cf.channel_id,
@@ -747,6 +790,48 @@ func (q *Queries) GetChannelFeaturesAndExtras(ctx context.Context, channelID int
 			&i.ExtraKey,
 			&i.Value,
 		); err != nil {
+			return nil, err
+		}
+		items = append(items, i)
+	}
+	if err := rows.Close(); err != nil {
+		return nil, err
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return items, nil
+}
+
+const getChannelFeaturesBatch = `-- name: GetChannelFeaturesBatch :many
+SELECT
+    channel_id,
+    feature_bit
+FROM graph_channel_features
+WHERE channel_id IN (/*SLICE:chan_ids*/?)
+ORDER BY channel_id, feature_bit
+`
+
+func (q *Queries) GetChannelFeaturesBatch(ctx context.Context, chanIds []int64) ([]GraphChannelFeature, error) {
+	query := getChannelFeaturesBatch
+	var queryParams []interface{}
+	if len(chanIds) > 0 {
+		for _, v := range chanIds {
+			queryParams = append(queryParams, v)
+		}
+		query = strings.Replace(query, "/*SLICE:chan_ids*/?", makeQueryParams(len(queryParams), len(chanIds)), 1)
+	} else {
+		query = strings.Replace(query, "/*SLICE:chan_ids*/?", "NULL", 1)
+	}
+	rows, err := q.db.QueryContext(ctx, query, queryParams...)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var items []GraphChannelFeature
+	for rows.Next() {
+		var i GraphChannelFeature
+		if err := rows.Scan(&i.ChannelID, &i.FeatureBit); err != nil {
 			return nil, err
 		}
 		items = append(items, i)
@@ -840,6 +925,55 @@ func (q *Queries) GetChannelPolicyExtraTypes(ctx context.Context, arg GetChannel
 			&i.Type,
 			&i.Value,
 		); err != nil {
+			return nil, err
+		}
+		items = append(items, i)
+	}
+	if err := rows.Close(); err != nil {
+		return nil, err
+	}
+	if err := rows.Err(); err != nil {
+		return nil, err
+	}
+	return items, nil
+}
+
+const getChannelPolicyExtraTypesBatch = `-- name: GetChannelPolicyExtraTypesBatch :many
+SELECT
+    channel_policy_id as policy_id,
+    type,
+    value
+FROM graph_channel_policy_extra_types
+WHERE channel_policy_id IN (/*SLICE:policy_ids*/?)
+ORDER BY channel_policy_id, type
+`
+
+type GetChannelPolicyExtraTypesBatchRow struct {
+	PolicyID int64
+	Type     int64
+	Value    []byte
+}
+
+func (q *Queries) GetChannelPolicyExtraTypesBatch(ctx context.Context, policyIds []int64) ([]GetChannelPolicyExtraTypesBatchRow, error) {
+	query := getChannelPolicyExtraTypesBatch
+	var queryParams []interface{}
+	if len(policyIds) > 0 {
+		for _, v := range policyIds {
+			queryParams = append(queryParams, v)
+		}
+		query = strings.Replace(query, "/*SLICE:policy_ids*/?", makeQueryParams(len(queryParams), len(policyIds)), 1)
+	} else {
+		query = strings.Replace(query, "/*SLICE:policy_ids*/?", "NULL", 1)
+	}
+	rows, err := q.db.QueryContext(ctx, query, queryParams...)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+	var items []GetChannelPolicyExtraTypesBatchRow
+	for rows.Next() {
+		var i GetChannelPolicyExtraTypesBatchRow
+		if err := rows.Scan(&i.PolicyID, &i.Type, &i.Value); err != nil {
 			return nil, err
 		}
 		items = append(items, i)