This is the third PR in the graph-SQL performance improvement series. This one is aimed at improving the performance of the `ForEachChannel` method which iterates through all the channels & policies in the graph DB. 
This is done by:
1) using batched pre-loading to fetch the features/extra-types for a set of channels & policies instead of one by one. 
2) manipulating the default pagination limit size

part of https://github.com/lightningnetwork/lnd/issues/9795

# Benchmark Results

## ForEachChannel Performance

| Database | Baseline (ns/op) | Batching (ns/op) | Batching % | Page Size (ns/op) | Page Size % | Overall % |
|----------|------------------|------------------|------------|-------------------|-------------|-----------|
| kvdb-sqlite | 290,980,425 | 288,054,600 | N/A | 280,527,775 | N/A | N/A |
| native-sqlite | 29,855,410,267 | 1,461,287,217 | +95.1% | 1,438,825,942 | +1.5% | +95.2% |
| kvdb-postgres | 242,501,225 | 240,550,842 | N/A | 243,166,692 | N/A | N/A |
| native-postgres | 23,764,840,392 | 799,722,842 | +96.6% | 536,877,517 | +32.9% | +97.7% |